Fix `OverrideAllocateMemory` for trimmed traces
When capturing a replay of a capture and `vkAllocateMemory` is called
with `VkMemoryOpaqueCaptureAddressAllocateInfo` already in the `pNext`
chain, the precedent behavior was to "add" another
`VkMemoryOpaqueCaptureAddressAllocateInfo` in front of it.

I think the best behavior is to check for the presence of this struct
in the `pNext` chain and, if present, not to touch anything. That is
what this commit does.

Change-Id: I743b95dcb73b4d6137a9604d111627c3d1c31b5a