Fix DescriptorUpdateTemplate usage
Fix DescriptorUpdateTemplate code was not fully complete.
Some of the code needed to be manually generated, and some of the
autogeneration code was not being generated properly.
This includes properly ordering the descriptor information when the
vkUpdateDescriptorSetWithTemplate API entrypoing was called.

Also resolved a few more pNext issues.

Clean up scripts and generate output in local scopes to reduce
stack memory overflow issues.