Remove redundant calls to VulkanCaptureManager::Get() (#1279)
The VulkanCaptureManager::Get() function is static and returns a pointer to the
current capture manager instance. While the compiler should inline these calls,
it may fail due to possible aliasing as the instance pointer is not const (but
should never change after setup).
It is better to call get once and explicitly put it in a local variable.

Only vkCreateInstance needs special handling, since it creates the
VulkanCaptureManager in the middle of the function.