Support for VK_EXT_frame_boundary
Add support for the frame boundary extension
VK_EXT_frame_boundary. This includes:
- `VkFrameBoundaryEXT` structures found in the pNext chains
of `vkSubmitInfo` and `vkSubmitInfo2` are now interpreted as
frame delimiters at capture time if `flags` member has the
`VK_FRAME_BOUNDARY_FRAME_END_BIT_EXT` bit set.
- The same structures in the same calls support screenshots
at replay time by interpreting `pImages` as the rendered
frames.
- `VK_EXT_frame_boundary` is now listed as a supported layer
device extension (programs that use the layer have access to
the extension even if the driver does not support it)

Extended support for this extension that is NOT implemented
by this commit could include:
- Interpretation of `VkFrameBoundaryEXT` structure when
found in `VkPresentInfoKHR` and `VkBindSparseInfo` (eg.
by considering the *real* render frames to be the ones
specified by the structures instead of the one presented)
- Support for `VkFrameBoundaryEXT` even when the
`VK_FRAME_BOUNDARY_FRAME_END_BIT_EXT` bit is not set (eg.
by saving the resources associated in a map/vector to be
taken into account when the real frame end is encountered)

Change-Id: I4615b490e46a108c58fd1ab2a07488a4ac8bd732