Add `--offscreen-swapchain-frame-boundary` option
This option inserts a `VkFrameBoundaryEXT` from `VK_EXT_frame_boundary`
into a command buffer submission called where `vkQueuePresentKHR` should
have been called by an offscreen swapchain.

By doing this, it is now possible to capture a trace replayed with the
option `--swapchain offscreen` AND to still have frame numbering and the
possibility to take screenshots.

This is very important for automated pipelines that run on a server
without WSI context and still need to take screenshots, or do
fast forwarding...

Change-Id: Iaa6252c014169af34b24eb719cfd3e23cd638c9f