Removed Empty Default Cases
It is good practice to avoid a default case that does nothing for
a switch case statement that switches on an enum so that the compiler
will complain if one of the enumerants is accidentally missed.
By extension it is good practice to always avoid such empty defaults
so as not to get in the habit of writing them and to save on lines
of code which do nothing.