Better handle for RT trimmed captures from optimize tool
The optimize tool was not taking under consideration ray tracing
features. As a result when applied on a trimmed capture containing such
calls all relative resources were ommited from the optimized tool
leading to corruption during replay.

One key element missing is the relation between top (tlas) and bottom
(blas) level acceleration structures. This relation is described when
the tlas is being build. However when capturing trimmed the tlas build
commands are likely not to be within the capture range and as a result
they are not dumped into the capture file. This makes it impossible for
the optimize tool to deduce from the used descriptors which blases it
should retain in the optimized file.

To solve this a new metacommand is introduced which describes the tlas
to blas dependency and is writted to the capture file during capture.
For each tlas that is built the new metacommand is written to the
capture file and it's essentially a list of all the blas ids the tlas
uses.