Implement toCpp virtual swapchain code
Implement the virtual swapchain path in the toCpp code.  This is
done by writing a swapchain_common header and source file which
handles the swapchain code.  By default, the virtual swapchain
code is enabled but can be disabled.  In generated code, this can
be done simply by modifying the define of "USE_VIRTUAL_SWAPCHAIN"
from 1 to 0.

It can also be disabled when generating the source by using the
"--captured-swapchain" option.