Add Wayland protocol dynamic loader generator
Add Python script `generate_wayland.py` that generates source
code from a Wayland XML protocol. This script has the same
role as the `wayland-scanner` tool, except that the sources
generated use the GFXReconstruct `WaylandLoader` instead of
directly using `libwayland-client.so` symbols.

The current state now is:
- If the machine that compiles the code does not support
Wayland (no `libwayland-client.so`) then the Wayland code
is simply not compiled and the executable will not support
Wayland (as done until now)
- If the machine that compiles the code supports Wayland
and the executable is ran on a machine that supports
Wayland, then the `libwayland-client.so` of the client
will be loaded using `dlopen` and the generated protocol
sources will initialize protocol constants (interfaces).
- If the machine that compiles the code supports Wayland
but the executable is ran on a machine that does not
support Wayland, then the WaylandLoader will remain
uninitialized and thus not call the generated protocol
sources and not attempt to initialize the protocol
constants.

This way, the binaries will be able no matter the state of
the support of Wayland on the machine that runs them.

Change-Id: I9b3878a71b74002859f4edda6d9a6fe180278a8f