Fix replay assert with imageless framebuffers
With the VK_KHR_imageless_framebuffer extension or the Vulkan 1.2
VkPhysicalDeviceVulkan12Features::imagelessFramebuffer feature, the
application provides image views to be used as framebuffer attachements
in vkCmdBeginRenderPass instead of during framebuffer creation. The
current code was only tracking attachments provided at framebuffer
creation and asserting when they didn't match the attachment count from
the render pass. This change will look for attachment image views either
from vkCreateFramebuffer or vkCmdBeginRenderPass, depending on whether
the framebuffer was created with the VK_FRAMEBUFFER_CREATE_IMAGELESS_BIT
flag. This issue was found by capturing and replaying ANGLE content,
which prefers to use imageless framebuffers if the feature is available.