Add utility for walking decoded pNext chains
There are now several places in the replay consumer that have the same
awkward loop to search decoded pNext chains for a particular struct.
This change adds a utility function to handle the search and cast the
result, along with a generated utility to automatically determine the
correct VkStructureType value to search for based on the meta struct type.