Fix VK_EXT_frame_boundary extension activation
In OverrideCreateInstance, the extension name was added to
filtered_extensions AFTER the vector size was saved in the create info
structure.

This is obviously an error (it still worked because this option is
mainly used when capturing the replay, and that the capture layer
ALWAYS support the extension even if not activated).

Change-Id: If043ca5511f5657a247f08d4faf02999a142fa1b