Integrate cpp code to latest (#1441)
* Samsung C++ code

This pulls in the changes made by Samsung engineers to get the
ball rolling with the "toCpp" transition.  Unfortunately, pulling
the source over rolled it all up into one commit, so in order
to not lose some history (and the credit) for their hard work,
we wanted to list the contributing engineers here:

  Andy Poole - a.poole@samsung.com
  Péter Gál - pgal.usz@partner.samsung.com
  Zsolt Borbély - zsborbely.u-szeged@partner.samsung.com
  Tamás Kéri - tk-szeged@partner.samsung.com
  Adrián Rávai

Special thanks to their hard work.

* Fix CPP output for Talos and VkCube

Modified scripts to support a newer Vulkan Header commit.
Cube and Talos required additional changes to work.
The code was hard-coded to think about only 1 GPU so recorded the last
memory types only which failed if the last GPUs was not used.
Additionally, I accidentally left out hte Android files from the CPP
pull over from before.

* Add per-handle nextIdx to make comparing easier

* Fix pNext chain CPP output

Fix the pNext chain CPP code output, it was using an incorrect
pointer.
Fix extra spaces in the generated CPP code.
Remove using "GetExistingStruct".  It didn't help much and it also
caused compilation issues since it would attempt to reuse the structure
elsewhere it wasn't defined.

This almost has a Wolfenstein2 trimmed capture replaying.

* Update copyrights to include Google and LunarG

* Fix DescriptorUpdateTemplate usage

Fix DescriptorUpdateTemplate code was not fully complete.
Some of the code needed to be manually generated, and some of the
autogeneration code was not being generated properly.
This includes properly ordering the descriptor information when the
vkUpdateDescriptorSetWithTemplate API entrypoing was called.

Also resolved a few more pNext issues.

Clean up scripts and generate output in local scopes to reduce
stack memory overflow issues.

* Applying coding standards

Start applying GFXR coding standards to the files.

* Make toCPP conditional in CMake

1) Make the build of the `ToCpp` source conditional based on the makefile
   flag `GFXRECON_TOCPP_SUPPORT`.  If this is set to `ON`, it will build
   the source.  Otherwise, it will not build the files.  This is because
   the software is considered to be in an Alpha state.

2) Make some more coding standards adjustments.

3) Modified argument handling in the tocpp `main.cpp` source file

4) Update the README.md in the `tocpp` tools folder to detail
   how to use the tool and what the current limitations are.

* Remove xxHash source

GFXReconstruct had a hash function already in place that wasn't
used.  It was only for returning 32-bit hash values, so it was
therefore improved to support 8-bit, 16-bit, 32-bit, 64-bit and
128-bit hash values.  Only use the 64-bit version for now in
place of the xxHash version.

* Increase memory, steps towards handling HW Bufs

* Get Android HW Buffer support compiling

This gets toCpp captures using Android Hardware buffers generated
source compiling.
It also fixes the scenario where an application called
`vkCreateInstance` more than once and enumerated physical devices
more than once.  The original code expected only one of each of
these calls.

* Fix ImportAndroidHWBuffer structure

Fix the VkImportAndroidHardwareBufferInfoANDROID structure buffer
so that it points to the correct Android HW buffer

* Fix imported semaphore support

Fix support for imported semaphores (i.e we ignore them on replay).
This change gets Aztec running on my Pixel device.
Also remove the global variable tracking which never worked properly
for me and other debug code that crept it's way in

* Add initial opaque memory support

Initial pass at attempting to implement opaque memory support.
Looking for example to test against, but for now it compiles and
doesn't interfere with the other tests that are run.

* Fix Aztec swapchain replay count

Fix the swapchain replay count for Aztec by checking the swapchain
requirements of the replay system.

* First step to custom swapchain code

First step to pulling over the virtual swapchain code changes into the
toCPP code path.

* Implement toCpp virtual swapchain code

Implement the virtual swapchain path in the toCpp code.  This is
done by writing a swapchain_common header and source file which
handles the swapchain code.  By default, the virtual swapchain
code is enabled but can be disabled.  In generated code, this can
be done simply by modifying the define of "USE_VIRTUAL_SWAPCHAIN"
from 1 to 0.

It can also be disabled when generating the source by using the
"--captured-swapchain" option.

* Add note in readme

* toCpp: Remove pre-processor step

The pre-processor step was unnecessary.  In addition, this change
brings it more in line with how the replay application works.

Finally, by making this change, I was able to get the resize
behavior working for desktop replay.

* toCpp: Remove resource tracker

The resource tracker was unnecessary and all handle tracking is actually
done by the VulkanCppConsumerBase itself.

* Add Android framebuffer boundary catch

* toCpp: Generator cleanup

Remove no longer needed post_ calls in generator

* toCpp: Fix a few minor issues

Moving to Ubuntu 22.04 for my system revealed new failures.  Fixed
them and a few unnecessary items in the struct generator python
script.

* toCpp: Update generated source

Update the generated source to the headers for the latest rebase.

* toCpp: Fix talos replay on RADV

Talos replay on RADV encountered a fault.  We weren't loading the
appropriate function and attempted to dereference a null-pointer.

---------

Co-authored-by: Mark Young <marky@lunarg.com>
Co-authored-by: bradgrantham-lunarg <50641407+bradgrantham-lunarg@users.noreply.github.com>