Using userfaultfd in page guard manager (#1161)
* Using userfaultfd in page guard manager

Introducing an alternative memory tracking mode which utilizes the
userfaultfd mechanism provided by the Linux kernel. This new mode should
provide an alternative for Linux and Android applications that are not
possible to be captured with the page_guard mode because they interfere
with the SIGSEGV handler. The new mode is enabled by seeting
GFXRECON_MEMORY_TRACKING_MODE/ebug.gfxrecon.memory_tracking_mode to
uffd.

* fix spelling

* match mem track mode language with other instances

* Delegating constructor

Converted the constructor with the default values into a delegating
constructor

* Disable uffd for macos

Also use uint64_t for thread ids instead of pid_t

* Remove UFFD_FEATURE_PAGEFAULT_FLAG_WP

* Temporarily use userfaultfd as default

---------

Co-authored-by: bradgrantham-lunarg <50641407+bradgrantham-lunarg@users.noreply.github.com>
Co-authored-by: ziga-lunarg <ziga@lunarg.com>