Clarify Vulkan-specific items (#1443)
* Vulkan Instance and Device functions (and tables) were only identified
by the Instance or Device label and not clearly identified as being
Vulkan-specific.  Add "Vulkan" to each to clarify their usage so
that they do not conflict with other Khronos-related API items.

* Update Vulkan-specific wrappers in the encoder to have a "Vulkan"
identifier to differentiate it from other Khronos API related items.

* Move a number of Vulkan-specific items into a few namespaces so that
it is clear which items are Vulkan in the future when other APIs
have been added into the standard layer.