Track AHB buffers with userfaultfd
Although original page guard memory tracking method (mprotect) is
tracking imported AHB buffers, the initial userfaultfd commit did not
add support for them.

With this commit imported AHB buffer are now tracked with the
userfaultfd method too.

However there is a caveat:
Userfaultfd requires that both base address and length of registered
memory regions to be paged aligned. Since in the case of imported
buffers we cannot back the regions with a shadow memory (we cannot
intercept their allocation and return a shadow pointer instead of the
original one), when registering the buffer to uffd we need to page align
both base address and length. This can effectively extend the tracking
to memory that is not inside the actual buffer and can lead to undefined
behavior. A warning message is printed if such a case is detected.