Fix "handle research" in code generation
The method `BaseGenerator.check_struct_pnext_handles` in
`base_generator.py` should return `True` if the specified struct
contains handles, handle pointers, or if a struct that contains handle
or handle pointers can be referenced through it (either by the `pNext`
chain or with a member that would be a struct or struct pointer).

Until now, to determine if a struct can reference such another struct,
the method was retrieving `valid_extension_structs` which query structs
that contains a field `structextends` with the checked struct.
The issue is that if struct `A` and struct `B` both extend struct `C`,
then struct `A` can be in the `pNext` chain of struct `B` (and
otherwise) which was not detected.

This commit proposes to solve this issue by including in the
`valid_extensions_structs` all the structs extending the parent struct
of the checked struct.

Change-Id: I1c75157d2f9076293a1f2e4c66926e2a75e7d048