Fix shadow semaphores used with VkSubmitInfo
A semaphore can be signaled with both vkAcquireNextImage and
with VkSubmitInfo pass to vkQueueSubmit.

If a null-swapchain/surface is signaling a semaphore we use the
shadow_signaled variable to track the state of the semaphore, instead of
actually signaling and waiting on it with vk commands.

This works if such a semaphore is only used with vkAcquireNextImageKHR
and vkQueuePresentKHR, but if the same semaphore is being used with
VkSubmitInfo passed to vkQueueSubmit the semaphore will never be
signaled, but will be waited on in vkQueueSubmit.

To fix this we add a variable signaled, which is used to track the
actual semaphore state when used in VkSubmitInfo.