openxr: Manually handle xrEnumerateSwapchainImages
Because of the way the xrEnumerateSwapchainImages uses the
XrSwapchainImagesBaseHeader struct in a direct array, we can't
dynamically allocate the array elements, but we have to allocate
it all up front and then read the correct data.  Because of this,
make a "Peak" function which looks ahead and determines the appropriate
structure type and then perform the allocate and read of the correct
structures.