openxr: Prevent holding of call lock over dispatch
Given that the capture layer might be reentrant when used for mulitple
API, (specifically OpenXR ATM), the call capture can't safely hold locks
when calling down-layer.  The DestoryLock is still in place for Destroy
calls, and while it hasn't proven reentrant on those paths to date, is a
potential (theoretical) risk for deadlock.