Make FileProcessor::ProcessBlocks common code reuseable
Any class extending FileProcessor does so to modify how blocks are processed.
The logic ProcessBlocks does not lend itself well to abstraction, so the best
alternative is to move as much code into the subfunctions as possible.

The offending code was frame delimiter code, which was properly copied into
subclasses, while logic to fixup old traces lacking frame delimiters was not.