Fix `vkSetDebugUtilsObjectNameEXT`
Fix `vkSetDebugUtilsObjectNameEXT` and `vkSetDebugUtilsObjectTagEXT`
for `VkDeviceMemory` objects when using rebind allocator.

When using rebind allocator, lazy allocation make it impossible to
have a valid Vulkan handle before binding memory. However, it remains
possible to set name and tag of the memory before binding it.

These informations should thus be stored in the allocator, and the
appropriate call emitted once the memory is actually allocated.

In addition, even without this lazy allocation issue, only a "fake"
handle is stored in `DeviceMemoryInfo`, making it impossible to use
even after binding time.

All of this is solved by this commit.

Change-Id: I8894f3111b2708fe6d8dc7d20a9204b4e9000c06