Enable an ability to mark api calls injected by replayer for tool layers to differentiate them from actual replay calls
This change enables the replayer to load the layers provided via
a replay option by name, separated by comma, and then mark the start and end of
the calls that are injected by replayer with the callbacks provided by the layers.

This change also attempts to mark some of the  known injected calls.

Change-Id: I03edc8bbd7874d79339ce56134d35349072e5e16