Fix noop functions address comparison
It sometimes is necessary to check if a vulkan call in a table is
noop or not.
Current prototypes of noop functions include the `static` keyword,
making address comparison impossible because the address changes
for each translation unit.

These functions are all one-liners with a simple call to a logging
message. I propose to replace the `static` keyword with `inline`.
The C++ standard specifies that inline functions have the same
address accross translation units.

Change-Id: If39e134d4d0cef2d818541b6b376eb26bf001880