Add codegen inout parameter support
Update the DX code generators to treat INOUT parameters as OUT
parameters. All INOUT parameters are pointers to integers that specify
and retrieve sizes for variable length arrays, where the input is the
size of the array and the output is the total number of items copied to
the array.