Custom decoder for CheckFeatureSupport data param
Adds a new DxFeatureDataDecoder type to encapsulate the
PointerDecoder<T> and StructPointerDecoder<T> types needed to decode
the CheckFeatureSupport pFeatureSupportData parameter in a generic type
that can be provided to DX consumers. Provides the same pointer decoder
interface as PointerDecoder/StructPointerDecoder for managing pointer
data, as well as functions for accessing the internal
PointerDecoder/StructPointerDecoder with a specific type. Also allows
the output allocation to be performed by the consumer instead of the
decoder, which matches the standard consumer behavior.