Fix ResizeBuffers replay with width/height = 0
When IDXGISwapChain::ResizeBuffers is called with a width or height
value that is equal to 0, that dimension will be resized to match the
dimension of the client area of the target window. For this case, replay
was resizing the window to have a width/height of 0, resulting in the
client area no longer being visible. This change specifies the
width/height of the window's client area to Window::SetSize, instead of
0, when width/height are 0.