Fix gfxrecon-info ResizeBuffers usage
When IDXGISwapChain::ResizeBuffers is called with a width or height
value that is equal to 0, that dimension will be resized to match the
dimension of the client area of the target window. When the stats
consumer used by gfxrecon-info processesed a ResizeBuffers call with
width and height equal to zero, it reported the swap chain size as 0x0.
The stats conusmer has been updated to ignore the ResizeBuffers width
and height with a value of zero when determining swap chain size.