Handle UpdateSubresource data pointer adjustment
When ID3D11DeviceContext::UpdateSubresource is called from a deferred
context with a driver that does not support command lists and the
pDstBox parameter is non-null and the pDstBox parameter specifies a
non-zero start offset for the update, the application is required to
subtract the offset from the pSrcData parameter. Capture and replay have
been updated to account for this adjustment. Note that this adjustment
is not required for ID3D11DeviceContext1::UpdateSubresource1.