Make IDXGISwapChain buffer cast conditional
At replay, the buffer returned by IDXGISwapChain::GetBuffer was always
cast to an ID3D12Resource type, which can lead to a crash when the
caller requests a different type.

With this change, the returned buffer is only processed as an
ID3D12Resource type when the caller specifically requests an
ID3D12Resource type.