dx12: Trim drawcalls
For writing a queue submit, it starts and stops in
PreQueueSubmit and PostQueueSubmit, as
GFXRECON_CAPTURE_QUEUE_SUBMITS. Since it only write
one queue submit, the target commandlists and
drawcalls will be saved in Dx12StateWriter.

For writing commandlists in Dx12StateWriter::WriteCommandListState,
it writes commandlists from the beginning to the target one,
and skip the reset of commandlists.

For writing drawcalls in Dx12StateWriter::WriteCommandListCommands,
if the commandlist is not target, it writes all commands.
If the commandlist it the target, it writes commands from the
beginning to the target drawcall, and skip the reset of commands,
and write EndRenderPass, EndQuery and Close.