quest_replay: Fix openxr loader usage
The OpenXR loader wasn't properly being built for the Quest replay
tool.  This is because the Android project requires all built source
to have a functioning CMakelists.txt file under the Android base tree.
So copy the CMakeLists.txt files out of the external/OpenXR-SDK folder
into the android/tools/quest_replay under an "external" folder and
modify them to proplery build the loader for the quest target.