Raytracing update: device address scanning
Device addresses now are being scanned in FillMemoryCommands at capture.
If a known device address is found in mapped memory, its location is
recorded in "FixDeviceAddress" metacommand inserted right before
the relevant FillMemoryCommand.

Since the scan is based on a simple value comparison there's a risk
a random memory area will be marked as a device address location.
To mitigate that risk additional capture option was introduced, allowing
the user to skip device address lookup in certain buffers based on their
usage.

Change-Id: Ie8c250c352364744b3c76e308b171abf57f55b0d