Fix D3D12 WriteToSubresource size calculation
The current WriteToSubresource size calculation may not work correctly
when either the DstSubresource parameter is greater than 0 or the
subresource is a texture with a block compressed format.

The size calculation was updated to handle these cases using texture
utility code that was originally added for D3D11, where it is used for
the D3D11 UpdateSubresource and WriteToSubresource size calculations.