Avoid opaque pointer copies in vulkan_struct_deep_copy routine
- do not copy 'external'-object-handles
- do not generate overload for VkPushDescriptorSetWithTemplateInfoKHR
- based on fix provided by Dustin Graves

When making deep copies of structures, avoid attempting to copy the
values referenced by opaque handles and opaque pointers. These are
generally pointers to a single value with the void* type, which cannot
be safely copied because the size of the referenced value is unknown.
For most cases, the referenced value is unneeded because capture treats
the address stored in these handles/pointers as external object IDs
that are encoded as 64-bit integer values.

Additionally omit VkPushDescriptorSetWithTemplateInfoKHR from code-gen,
as it requires knowledge about the template's layout,
which is not available in the routine's scope.