Introducing SetInjectedCommandCallbacks().
This callback can be called by an external layer/library in order to
register 2 callbacks. One that gfxrecon-replay will call each time
it starts making API calls that are not part of the capture file,
and the other when it is done.

The callback can be detected with dlsym/GetProcAddress