Fix an error for parsing shader code
Certain optimizations may remove SPIR-V instructions such as OpName,
which can lead to a nullptr name in the spirv_reflect
SpvReflectDescriptorBinding struct. This occurs because spirv_reflect
sets the name by keeping track of these SPIR-V instructions. The commit
addresses an issue in handling this scenario when parsing the shader code.