Fix an error for VkSurfaceFullScreenExclusiveWin32InfoEXT handling
When generating deepcopy sorce code for
VkSurfaceFullScreenExclusiveWin32InfoEXT, the member "hmonitor" should be
treated as a non-pointer member because it is an opaque pointer defined
by the Windows platform. For a valid hmonitor, it is guaranteed that it is
not a null pointer, but it may not access the pointed memory. Therefore,
we should not generate source code to copy its pointed data. The commit
fixed a capture crash due to access violation.

Change-Id: I9f5f49888bb962df05a39a78a486a3c69d5dd566