Fix DX11 OMSetRTsAndUAVs encoding
When the ID3D11DeviceContext::OMSetRenderTargetsAndUnorderdAccessViews
NumUAVs parameter is equal to D3D11_KEEP_UNORDERED_ACCESS_VIEWS or
NumRTVs is equal to D3D11_KEEP_RENDER_TARGETS_AND_DEPTH_STENCIL, the
associated RTV and UAV array parameters should be ignored for encoding
because they may specify addresses that are not safe to access. This
requires custom encoder and replay implementations to ensure the arrays
are not dereferenced at capture and that -1 (the value of the KEEP
macros) is not used as the array size for decoder memory allocation and
handle mapping operations.