openxr: Unknown events in capture but not replay
Some unknown events were triggered in some OpenXR captures on the
Meta Quest headset, but are not showing up during my replay run.
So, if the event is not known by GFXR and it's not handled, just
continue onward as if nothing happened, but keep track of the event
in case it shows up at some point in the future.