Adjust output array sizes for D3D12 replay (#1785)
* Allocate output data for replay

Update code generation to allocate output data for output pointers.

* Add codegen inout parameter support

Update the DX code generators to treat INOUT parameters as OUT
parameters. All INOUT parameters are pointers to integers that specify
and retrieve sizes for variable length arrays, where the input is the
size of the array and the output is the total number of items copied to
the array.

* Gen AddObject for non HRESULT case

With D3D12, all functions that retrieve a COM object also return
HRESULT. With D3D11, there are some functions that retrieve a COM object
while returning void. The code generator has been updated to handle this
case, generating calls to AddObject() for the COM objects retrieved by
functions that don't return an HRESULT value.

* Support D3D 'result_bytebuffer' case

D3D output parameters with a SAL labl including the text
'result_bytebuffer' and a size appear similar to arrays of bytes but are
really a pointer to memory allocated internally by the implementation
with the specified size. For this case, the following adjustments are
made for capture and replay, where capture will continue to record the
content of the retrieved memory and replay will allocate and retrieve a
single pointer to the memory.
- Update encoding for this case, where the type is void**, to
  dereference the pointer when invoking EncodeVoidArray.
- Update replay to decode the array of bytes, but allocate a single
  pointer parameter for API call replay, using PointerDecoder<uint8_,
  void*> to represent the parameter.

* Adjust variable length output array sizes

Update D3D12 replay code generator to apply the same variable sized
output array adjustment that is used for Vulkan. For API calls that have
an optional output array parameter and a pointer to an integer size
parameter, when the array parameter is NULL, store the retrieved size in
the calling objects info struct, and then when the API call is made with
a non-NULL array parameter, specify the size stored in the info struct
when allocating the output array.

* Prevent array size null pointer dereference

Update code generation to add a null check when dereferencing
a pointer to an array size when encoding array data during capture and
allocating array memory during replay. Added for cases where both the
variable length array parameter and its associated size parameter are
optional and are allowed to be null.