Copy AHB without CPU read mask
Import non-CPU readable Android Hardware Buffers with a VkImage. Use a
Ycbcr conversion sampler to get data from external unknown format to
RGBA8. Finally, read the data from GPU to host for a FillImageMemory
command.

Importing external memory with external format is not supported at
replay time. The capture stored data in RGBA8_UNORM.
The content of the Android Hardware Buffer is restored with a Vulkan
image which imports the AHB memory, then it's filled reusing
functionalities of the resource initializer.
Finally, both image views and samplers are patched to remove any
reference to Android hardware buffers and external memory formats.

Do not write the state of image memory backed by an AHB as the AHB has
already been processed. At replay time the content of the image will
clearly correspond to the content of the AHB.

If image is not VK_NULL_HANDLE and the memory is not an imported AHB or
an imported QNX Screen buffer, VkMemoryAllocateInfo::allocationSize must
equal the VkMemoryRequirements::size of the image.

If memory is importing an Android Hardware Buffer, its allocation size
should be equal to the AHB properties allocation size.