dx12 option: GFXRECON_CAPTURE_DRAW_CALLS  (#1680)
* dx12 option: GFXRECON_CAPTURE_DRAWCALLS

Specify one index or a range indices drawacalls(include dispatch)
based on a ExecuteCommandList index and a CommandList index to
capture. The index is 1-based. The args are one submit index,
one command index, one or a range indices of draw calls, one
or a range indices of bundle drawcalls(option), like "1,1,1" or
"1,1,1-2" or "1,1,1,1-2". The forth arg is an option for bundle
case. If the the 3rd arg is a bundle commandlist, but it doesn't
set the 4th arg, it will set 1 as default. Default is: Empty string
(all drawcalls are captured).

* Not unwrapped objects for override functions

Override functions need wrapped objects and wrappers.

* Split commandlist into three for target drawcalls

It does something similar to dump resources,
but a bit different. Trimming drawcalls is not like
dump resources that could browse the captured file
to find target commandlist and drawcall before replay.

For trimming drawcalls, it doesn't know the
target commandlist until running the target
ExecuteCommandLists. In this case, it assumes every
commandlist could be the target commandlist. Every
commandlist has three splitted commandlists and
add command to the splitted commandlists.

It also can't know which drawcall is the target in
advance, commandlist has a drawcall_count to count it.
Splitted commandlists will add some redundant commands
since it couldn't know the target drawcall in advance,
But it shouldn't affect the result and performance.

* Active and deactive trimming drawcalls

* Trim drawcalls chages to 0-based

* bugfix for bundle

* Rename drawcall to draw call and Bundary to Boundary

* Track swapchain buffers

For copy swapchain buffers:
1. The queue has to been swapchain's queue.
2. One ExecuteCommandLists could work for only one swapchain buffer.
3. The current back buffer index has to match the swapchain buffer.
4. After ExecuteCommandLists, the current back buffer index has to back init.

* Swapchain's queue to write its buffer

* Check bounds for trim draw calls

* Track extra ExecuteCommandLists

The original code skips writing and tracking extra ExecuteCommandLists.
But the tracking is necessary for the correct resource states.
The extra ExecuteCommandLists are for splitted commandlists of trimming drawcalls.

* QueryInterface to reinterpret_cast

* Move before extra ExecuteCommandLists

Move before extra ExecuteCommandLists from PreProcess_ID3D12CommandQueue_ExecuteCommandLists to OverrideID3D12CommandQueue_ExecuteCommandLists.

All three extra ExecuteCommandLists are in OverrideID3D12CommandQueue_ExecuteCommandLists to make the code simpler.

* Refactor trim draw calls

Move code from Override functions to individual functions

* Print warning if using DATA_STATIC

DATA_STATIC could cause error for splitted commandlists. DATA_STATIC_WHILE_SET_AT_EXECUTE is better. Both trimming draw calls on capture and dump resources on replay have this issue.

* small changes

1. Remove using QueryInterface

2. More check before HasSplitCommandLists

3. Move ++list_info->draw_call_count to TrackCommandExecution
   In order to make counting draw_call is after GetCommandListsForTrimDrawCalls
   , also move Encode_XXXX() to after GetCommandListsForTrimDrawCalls.

* Add lock on active/deactive trim draw calls

* bugfix: break loop when it found the target