dx12 codgen: Break depedence upon Vulkan/Khronos
Remove the hard-coded dependency of the DX12 GFXR codegen upon our
Vulkan and the Khronos registry edit functionality (which the DX
side really doesn't use).
This includes removing the dependency on the "base_generators"
directory as well.
NOTE: This did require copying a functionality (like write) from
the Khronos source to maintain consistency (so I added the Khronos
copyright notice to the dx12_generators/base_generator.py file).

The reason behind this is there are big changes coming to the Vulkan
side in preparation of supporting other Khronos APIs and these changes
could break DX12.