codegen: Remove feature break
Remove the feature break because when we move everything the breaks
won't be in the same place.  This way the changes are done prior to
the move to make sure nothing important changes.
Plus, eventually, we'll remove the "process_cmds", "process_structs"
and "feature_break" variables since they are not really useful and
are often mis-used.