openxr: Update xrPollEvent to handle missing events
Update the replay logic to handle waiting for missing events.
If the event is a known important event, we will always wait for it.
If the event is not important, or unknown, every time we miss receiving
the event, we will wait less next time.  We will always wait at least
once for every event, but if it's an event that we never receive,
we will no longer than one xrPollEvent query for it.