Fix IDXGISwapChain::ResizeTargets capture
IDXGISwapChain::ResizeTarget will resize the window associated with the
swapchain. If the window has an event callback, the callback will be
invoked for a resize event before ResizeTarget returns. Any API calls
made by the callback will not be processed for capture due to the
current scope count, resulting in a capture file that will not replay
correctly due to missing API calls. The scope count needs to be
decremented prior to the call and resotored after the call to ensure
that API calls made by the callback are captured.