codegen: Add skip struct suppport
Add skip to enum->json body

I also modified the if check when adding the skip to continue
instead of adding another "and not".  This cleans up the
logic and shifts code to the left.

Rename enum to flag in generator

And ran yapf

Add skip to enum->json header

Add skip to each "if" and modify the if checks to add a continue to
move code to the left.

add skip to enum->string header

Also update if statement iwht skip and modified it to continue
allowing remaining code to shift left.

Add skip to enum->string body

Also modify if check to perform a continue allow code to shift
to the left.

Add ApiData flags