Fix D3D Texture3D size calculation
With 3D textures, applications will sometimes overlap slice data by
specifying a depth pitch value that is smaller than the row pitch value,
eg. RowPitch=1024 and DepthPitch=64, indicating that the next slice
starts 64 bytes into the current slice. To handle this case, we start
by calculating the size of the first slice and then add ((Slices - 1) *
DepthPitch) to it. When the slices aren't overlapped, this should
simplify to (Slices * DepthPitch). When they do overlap, it should
evaluate to (Height * RowPitch) + ((Slices - 1) * DepthPitch), which
produces the size of a complete slice plus the size of the overlapped
data.