codegen: Refactor codegen for consumers
* Move consumer ext struct func gen

Move the consumer replay base definition of the extended structure
functions into the khronos scripts.

* Move replay override loading to Khronos

Move the replay override mechanism into the khronos_base_generator
so that it can be shared with other APIs.

* Port consumer body to Khronos

* Update handle processing in consumer replay

* Update replay consumer body

Update the replay consumer body generator to remove more Vulkan
functionality from the Khronos base items.

NOTE: This adds 3 new lines in the generated file which based on the
comments in the Python should have been there from the beginning.

* Move json consumer body to Khronos

* Add khronos logic for instance and device items

* More instance/device cleanup for consumer replay

* Update khronos replay consumer

Move the functionality for the Vulkan replay consumer body
generator into the Khronos base.  This also moved several support
functions referenced in the Khronos base to the base instead of
relying on functions defined in the child class.