dx12 dump: Check duplicated notes (#1889)
* Check duplicated notes

Since of "before" and "after", the infos are written twice. But the
infos paths are the same, so the second time would override the first
time. It doesn't duplicate. But notes are array. The second time won't
override the first time. It causes duplicated data. In this case, it
needs to check if the note is duplicated before writting.

* bugfix: end is the last data not out of the data