codegen: DRY codegen, prep for multi-api
* Move the dispatch table header generator to the Khronos generator

so that it can be shared with other APIs.

* Rename Vulkan Base classes to VulkanBase

The file, the Base and the Base Options class are all renamed.

* Move soon to be common method to khronos base

Several VulkanBase method are also present in OpenXrBase.  Move these to
Khronos base for future DRYness.

* DRY begin/end file output

Centralize the output of all includes and namespaces, by making them
genOpts data driven