Remove retired swapchain images immediately
The implementation is allowed to destroy and reuse handles for non-acquired
swapchain images as soon as the parent swapchain is retired. Update the handle
tracking logic to treat these handles as destroyed, so that if they are reused,
we don't get collisions in the handle tables.

Android is particularly sensitive to this as it will often immediately reuse
image handles from retired swapchains, even before the retired swapchain is
destroyed.