Set replay_options: enable_vulkan and enable_d3d12
enable_vulkan and enable_d3d12 were always true.

If dump resources is for vulkan, we don't want Dx12PreProcessConsumer to
work for it.

Check if replay is vulkan or dx12 in VulkanPreProcessConsumer and
Dx12PreProcessConsumer. And then return to replay_options: enable_vulkan
and enable_d3d12.

In this case, PreProcessConsumer has two tasks. It always needs to run
PreProcessConsumer. It needs to check replay that is vulkan or dx12 at
least.