Fix Android multi-window surface recreation
This commit fixes two bugs.

The first is a simple forgotten call to `ANativeWindow_release` on
window destruction.

The second is a synchronization issue in the Java code:
During window recreation, a destroy call and a create call closely
follow each other. Each of these calls is sent to the UI thread,
that will then trigger the callbacks `surfaceDestroyed` and
`surfaceCreated` later.
If the call to `addNewView` happens before the UI thread calls
`surfaceDestroyed`, then `mSurface` is set to the destroyed
surface (which is not null), which allows `addNewView` to return
with the incorrect surface set, which results in a failed
surface recreation.
I propose to solve this bug by simply not setting `mSurface` in
`surfaceDestroyed` and `surfaceChanged` callbacks, as `mSurface`
is exclusively used to detect newly created surfaces.

Change-Id: I766a56272dbe53606d86bb6b95e31e5a508a31c9