Track & Sanitize usage of buffer references in pipelines (#2040)
The goal of this PR is to allow usage of buffer-references in shaders (SPV_KHR_physical_storage_buffer)
when using the -m rebind option.

This aligns with previous work for portable raytracing and follows identical principles.

We already parsed spirv shader-code for usage of SPV_KHR_physical_storage_buffer.
here we start using the extracted offsets and locations and derive information about actual buffers
and buffer-addresses during runtime, when the pipeline is used.

We basically collect an array of gpu-addresses, containing locations of stale addresses.
Those can be replaced with an injected compute-dispatch.
In case this is not possible (inside a renderpass), the replacer/sanitizer-pass is postponed until queue-submit.
(and syncing via fence there, 'could' be done via semaphore)

this allows successful replays of various cases which would previously crash.

Note: this is sufficient for base-addresses of buffers, 
but will require additional work to also support offset-addresses.

also contains:
-  improved spirv-parser with cap-check, forward spirv-reflect pass, then our backward custom parse.
- unit test for SpirVParsingUtil
- fixed holes in buffer/acceleration-structure tracking
- use vkSetDebugUtilsObjectNameEXT to annotate names of shadow-resources in VulkanAddressReplacer
- switch internal hash-function from murmur32 to xxhash32. less code and faster, same results for our purpose